fixup state table schema
