+++ crates/rooch-indexer/src/models/mod.rs
@@ -2,6 +2,5 @@
 // SPDX-License-Identifier: Apache-2.0
 
 pub mod events;
-pub mod global_states;
-pub mod leaf_states;
+pub mod states;
 pub mod transactions;
